+++ main.struct.go
@@ -7,7 +7,7 @@ import (
 )
 
 //----- Constants -----
-const version = "1.2.5"
+const version = "1.2.6"
 const appServiceManager = "com.hornbill.servicemanager"
 
 //----- Variables -----